+++ Documentation/oven/install-yang.sh
@@ -1,0 +1,1 @@
+sysrepoctl --install --yang=oven.yang