+++ clients/run-test.sh
@@ -1,38 +1,130 @@
 #!/bin/bash
 
-overall=0
+testtype="all"
+if [ "$1" = "unit" ]
+then
+  testtype="unit"
+fi
 
-echo ""
-echo "LINT Checks.."
+set pipefail -euo
+
+printf "\n\e[1;33mLint checks.....\e[0m\n"
 pycodestyle ./
 if [ $? != 0 ]
 then
+  printf "\n\e[1;31Lint checks.. Failed\e[0m\n"
   exit 1;
+else
+  printf "\n\e[1;32mLint checks.. Passed\e[0m\n"
+fi
+
+printf "\n\e[1;33mComplexity checks.....\e[0m\n"
+# xenon . -a C -m C -b C
+xenon . -a C -m C -b C 2>/dev/stdout| grep -v '_find_best_number_type" has a rank of D' | grep ERROR
+if [ $? == 0 ]
+then
+  printf "\n\e[1;31mComplexity checkss.. Failed\e[0m\n"
+  exit 1;
+else
+  printf "\n\e[1;32mComplexity checks.. Passed\e[0m\n"
 fi
 
 
-echo ""
-echo "Unit tests.. (Core Set)"
-python3 -m unittest discover test/unitcore
+if [ -d htmlcov ]
+then
+  rm -fr htmlcov
+fi
+if [ -d htmlcov-unit ]
+then
+  rm -fr htmlcov-unit
+fi
+if [ -d htmlcov-unitcore ]
+then
+  rm -fr htmlcov-unitcore
+fi
+if [ -d htmlcov-integration ]
+then
+  rm -fr htmlcov-integration
+fi
+
+function combine_reports {
+  if [ $1 = "1" ]
+  then
+    mv htmlcov htmlcov-unitcore
+    mv .coverage htmlcov-unitcore
+    return
+  fi
+
+  if [ $2 = "1" ]
+  then
+    mv htmlcov htmlcov-unit
+    mv .coverage htmlcov-unit
+    coverage combine htmlcov-unitcore/.coverage htmlcov-unit/.coverage
+    coverage html
+  fi
+
+  if [ $3 = "1" ]
+  then
+    mv htmlcov htmlcov-integration
+    mv .coverage htmlcov-integration
+    coverage combine htmlcov-unitcore/.coverage htmlcov-unit/.coverage htmlcov-integration/.coverage
+    coverage html
+  fi
+}
+
+printf "\n\e[1;33mUnit Tests (Core Set).....\e[0m\n"
+nose2 -s test/unitcore -t . -v --with-coverage --coverage-config .coveragerc --coverage-report html
 if [ $? != 0 ]
 then
+  printf "\n\e[1;31mUnit tests.. (Core Set) Failed\e[0m\n"
+  combine_reports 0 0 0 0
   exit 1;
+else
+  printf "\n\e[1;32mUnit tests.. (Core Set) Passed\e[0m\n"
+  combine_reports 1 0 0 0
 fi
 
 
-echo ""
-echo "Unit tests.. (Extended Set)"
-python3 -m unittest discover test/unit
+printf "\n\e[1;33mUnit Tests (Extended Set).....\e[0m\n"
+nose2 -s test/unit -t . -v --with-coverage --coverage-config .coveragerc --coverage-report html
 if [ $? != 0 ]
 then
+  printf "\n\e[1;31mUnit tests.. (Extended Set) Failed\e[0m\n"
+  combine_reports 0 0 0 0
   exit 1;
+else
+  printf "\n\e[1;32mUnit tests.. (Extended Set) Passed\e[0m\n"
+  combine_reports 0 1 0 0
 fi
 
-
-echo ""
-echo "Integration tests.."
-python3 -m unittest discover test/integration
-if [ $? != 0 ]
+if [ -f "/.dockerenv" ]
 then
-  exit 1;
+  printf "\n\e[1;33mIntegration Tests.....\e[0m\n"
+else
+  printf "\n\e[0;33mNot running inside docker - skipping integration tests.\e[0m\n"
+  testtype="unit"
 fi
+
+if [ "$testtype" = "unit" ]
+then
+  printf "\n\e[0;33mRequested to only run unit tests\e[0m\n"
+fi
+
+if [ "$testtype" = "all" ]
+then
+  nose2 -s test/integration -t . -v --with-coverage --coverage-config .coveragerc --coverage-report html
+  if [ $? != 0 ]
+  then
+    printf "\n\e[1;31mIntegration tests.. (Extended Set) Failed\e[0m\n"
+    combine_reports 0 0 0 0
+    exit 1;
+  else:
+    printf "\n\e[1;32mIntegration tests.. (Extended Set) Passed\e[0m\n"
+    combine_reports 0 0 1 0
+  fi
+fi
+
+printf "\n\e[1;33mCoverage Report.....\e[0m\n"
+coverage report
+
+printf "\n\n\e[1;32mAll tests passed!!\e[0m\n"