+++ CHANGELOG.md
@@ -58,3 +58,10 @@
 # v0.0.4 - Bugfixes
 
    - FEATURE: Choices/Cases applied via a template now calculate the correct data path.
+
+# DEVEL version
+
+  - Sysrepo version bumped from 0.7.7 to 660552222ee6376efa560d9bcc7b832886ff460a
+  - Implemented dump_xpaths() to sysrepo data abstraction layer.
+  - Implement to_xmlstr in TemplateNinja to convert a list of XPATH's to XML.
+  - Leaf-List entrires no-longer duplicated in the stub store.